app/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
pytest
httpx
pwdlib[argon2]